Pod: allow empty config value
Previously a Pod block such as

  =begin pod :config('')
  =end pod

or

  =begin pod :config[]
  =end pod

would throw a compiler error (coercing NQPMu to Str). This is because,
first, an empty quoted string was treated like nothing when compiling
the list of values of the :config key above, and second, an empty value
list was unexpected on the caller site.

This commit makes empty quoted strings recognized and empty value lists
supported, so that

  =begin pod :x('') :y( )
  =end pod

parses successfully and gives

  dd $=pod.head.config
  #= Hash %!config = {:x(""), :y($[])}

Fixes #3098 and closes #3101