Make subset behave depending on what language revision created it
For the purpose of supporting language version dependent functionality a
new metamodel role is being introduced:
`Perl6::Metamodel::LanguageRevision`. It is in turn consumes
`Perl6::Metamodel::Versioning`.

SubsetHOW records the compiler version which created a subset type
object and then later checks against the version in order to determine
if old or new behaviour is expected.