Don't recompile a dependency if another process did it for us
When a direct dependency of the current compilation unit is outdated by
a change of one of the transitive dependencies, after taking the lock
for the precomp store, we now check if the dependency became up-to-date
while we were waiting for a lock. It may be that another process did the
work for us.

This helps especially with running tests of a suite of modules in
parallel. When changing a base module, multiple test processes would
often precompile the same high-level module and even wait for one
another.