Finalize implementation of Grammar.parse returning Failure
Based on rakudo/rakudo#3138 because Grammar class is referencing
HLL::Compiler class which is only available when Perl6::BOOTSTRAP is
preloaded for a CORE. Since previously `d` and `e` `CORE`s were compiled
as decent modules and didn't have their own `BOOTSTRAP`, use of `6.e`
version of grammar wasn't possible due to unresolved symbol.

This branch is a further development of rakudo/rakudo#3103.