Use the NativeCall repr's new get_int to avoid redundancy
Keeping the setup state in an attribute is redundant, since at the VM level we
know perfectly well, if the native call site was initialized or not. Even more
the latter may change due to repossession. Avoid any discrepancies by just
asking the VM for the state instead of keeping track of it ourselves.