Fix inconsistent state of NativeCall subs after repossession
When multiple modules depend on the same NativeCall using dependency, and some
of them call native subs during compilation we could end up with an
inconsistent state of the native sub. Both the sub's type and the sub itself
got modified during setup. The type's invocation protocol was changed and the
sub's code replaced. It could happen however that only one of them got
repossessed and we ended up with a changed invocation protocol but the original
code body.

Fix by getting rid of CALL-ME entirely. Instead, replace the routine's body
with what CALL-ME used to do. This way the invocation protocol will always stay
the same and we end up with an either fully set up routine or one that will run
the setup code. This also means that cloning the method (for closures) will no
longer cause issues as we don't modify the type anymore. So clones of native
subs will also get the speedups of optimized bodies.