Fix native subs declared in BEGIN blocks and role bodies
When a native sub was declared within a BEGIN block or a role body, the
replacement code body did not stick but was replaced with the original stub
when the containing block was compiled dynamically.

Fix by removing the code object from the map of objects needing fixup.
Fixes GH #3235 and #3243