Fix leaking an open file handle when re-precompiling
When checking if another process has precompiled the exact module that we need
precompiled while we were waiting for the lock on the precomp repository, we
opened the unit but never closed it. This would trip up Windows which refuses
to rename over an open file. Fix by ensuring that we close the unit when we're
done with our checks.

Hopefully fixes GH #3170