Improve error msgs: invalid open bracket, declarator blocks
Modified embedded declarator block token to throw a better exception
message.

Added another line to an existing exception message to direct the user
to the possibility of an improper open bracket for a declarator block.

Added three new exception handlers for the three declarator block
types. Not all are used yet but are in place wnen the error detection
improves in the future.

Removed old, invalid TODO lines, added more TODO and NOTE info.

This fix addresses Rakudo issues #3144 and #3145, but someone else
will have to close them if they think it is justified with the current
commit.

Additional work needed: I believe a better and more explicit message for
the exact type of declarator block will require modification of the
appropriate "quibble(self.slang_grammar('Quote')" chunk in Grammar.nqp
in a future fix.