Make PROCESS and GLOBAL set dynamic flag on manually created containers
When a container created via `ASSIGN-KEY` it's dynamic flag is set to
true to maintain consistency with other dynamic symbols. I.e.:

```
PROCESS::<$foo> = 1; $*foo.VAR.dynamic; # True
```

Support is done via:

- Introduction of `DynamicStash` class which is a child of `Stash`.

  The leaves the original Stash unchanged and doesn't impact it's
performance. Because assignments like in the example above are rare
events the price of setting the dynamic flag is neglectible.

- Support for dynamic stashes by MOP