Store resources files by their original name if safe
Windows expects DLL files to have their original name. This required
module authors to use workarounds like copying the resource file to a
temporary directory.

For the common case of file names using ASCII characters only, we store
the files using their original name. To avoid collissions, we create a
directory for the file using the generated file-id.

This is backwards compatible to already installed modules as we store
the full path relative to resources in the meta data.