Add support for implicit signature of { ... }
The .perl of parameters did not account for the case where a parameter
was taking the $_ of the outer scope.  This is now simulated by adding
" = OUTER<$_>" in that case.  Please note however that this is a bit of
a kludge, as this will *not* roundtrip correctly.  But it will at least
convey the difference with a block that had a specific signature that
was otherwise identical.