Fix Stash'es clashing after deserialization
When two modules are precompiled independently but then loaded by the
same process it is possible that if they share common name prefix like
in, say, `A::B::C::D1` and `A::B::C::D2` then the module loaded second
would get its Stash on `A` overriden by the module loaded first. This
would effectively result in some symbols in `A::B::...` hierarchy to be
lost to the code importing the modules. This behavior is caused by the
fact that when second module is being loaded by MoarVM from its
precompilation, symbols from parent stashes are being resolved to the
existing type objects from the first module. I.e., if `D1` is loaded
first, then upon loading `D2` it's `A` package would be resolved to the
`A` implicitly created earlier for `D1`. And that package would bear a
`Stash` object which contains `D1` but knows nothing about `D2`.

The problem is being fixed by preserving a `Stash` structure in a form
which wouldn't lost child objects after MoarVM resolves package symbols.
Currently it is done by creating a nested list of items containing
symbol name, symbol value and, optionally, subtable built from symbol's
`Stash` if it's a package. The structure is then stored in `Stash`
attribute `@!snapshot` before serialization. Then symbols from it are
merged back into the `Stash` object after deserialization.

It should also be noted that because `Stash` is Raku's level entity, it
makes sense to take care of it in Raku. Thus the solution of `Stash`
taking care of itself.

Fixes rakudo/rakudo#3075