Finalize implementation of relaxed FQN method calls
Method is now being looked up upon the type object to which the calling
code belongs. In case of roles the lookup will be performed on role's
concretization to have cases of parameterized roles handled correctly.

If lookup on the type object fails then it falls back to the object
instance type. I.e. it is performed against `self.WHAT`.