Fix relocatability of bundled native libraries
Native libraries (and other resource files) bundled with installed modules are
supposed to be relocatable, i.e. their path is always relative to the current
path of the repository. This is especially important for packaging as the
module will get installed into a temporary build directory using the Staging
repository. The generated files will be moved to their final location when
installed.

Serializing the current path of the library as part of the NativeCall site
would prohibit this relocatability, so block serialization for bundled
resources during precompilation instead.