An attempt at rewriting the Env initialization logic
Instead of creating a Hash at each startup with the val()ed values
of the outer environment, simply clone the result of nqp::getenvhash
and put this in a subclass of Hash, and provide the val() logic on
demand.

This proved quite an exercise in yak-shaving, and the result was really
inconclusive for yours truly.  On top of that, it appears to tickle a
precompilation bug that only shows up when doing "make spectest" with
a high number of TEST_JOBS.  Rather than just throwing away this code,
I thought I'd keep it in a branch, just in case someone could make use
of this in the future.