Fix "expected IO::Handle:D but got IO::Handle" in parallel test runs
When a process tries to precompile a dependency and finds that a different
process already does (e.g. test files run in parallel), it closes the
precompilation unit as to not leak the open file handle. However when it tries
to use the generated precomp file, it didn't re-open the file handle, because
the $!initialized flag on the precomp unit stayed on True.
Fix by re-setting $!initialized, so we actually try to open the file again.