Re-imagine val() handling
This replaces the old val() handling by a completely re-imagined version that
caters to the way that Raku is being optimized.  The goal was to reach an
overall speed increase of 2x, but this was not achieved.  But this commit makes
val handling more correct, and about 2x faster in the most common case, namely
that *no* number is encountered.  Which helps at every startup, as each %*ENV
entry is run through val().  This made the spectest use about 12 CPU seconds
less on yours truly's machine, which is only .3% on the total, but still.

So what are the perks of this change?

- Addition of the Str.val method
  Since "self" is sorta optimized especially in Raku, it made sense to create
  a method in the Str class doing "val" processing, allowing the string to be
  passed around implicitely on helper methods.
- complete merge of val() and Str.Numeric logic
  Both use the Str!val private method, but handle failures differently.
- Failure of parsing creates an internal "cheap" Failed object
  This is much cheaper than a Failure/Exception combo.  Especially as val()
  processing doesn't need the Failure object.
- Additional features implemented:
  - support for ∞ as a synonym for Inf
  - support for +i and -i as synonym for +1i and -1i
  - every location that handles - (hyphen) also handles − (minus)
  - support for Blob literals: :16{ aa bb } is the same as Blob.new(0xaa,0xbb)
- Fixed several precedence issues
- Performance differences:
  - Only 60% of allocations needed
  - Integers about 2x as fast
  - Non-numeric strings about 1.8x as fast
  - Empty string about 3x as fast
  - Other expressions a few percent faster or slower
- Testing:
  S32-32str/val.t contains *extensive* testing of all old and new features
  and has been added to t/spectest.data.6.c
- Other changes:
  - locations that used val() in the core, now call the Str.val method

TODO:
It is my intent to change the handling of the Raku grammar and its associated
actions, to call the Str.val logic during compilation, to ensure that all
val()-like processing is identical, whether it is done during compilation using
the < > syntax, or it is used for numeric literal parsing such as :16<a b c>.
This should replace a significant part of the Raku grammar / action logic by
Raku source code (albeit heavy using nqp:: ops).