Make `take foo, bar` about 60% faster
By not using the helper THROW or RETURN-LIST subs, but instead do all
of the stuff inside `take`.  Which could be made a lot simpler, as that
candidate can only be called if there is more than one parameter, so
the whole check for number of elements was not necessary.