Fix a problem when attribute object doesn't have $!package set for roles
The fix is rather simple: insted of setting `$!package` to generic
`$?CLASS` set it directly to $*PACKAGE as it contains the role/class
being currently compiled and which actually owns the attribute. This
approach requries additional fixup later, when the attribute is getting
actually composed into a class. At this point `$!package` must be set to
the class typeobject. The previous commit guarantees that the action is
performed upon a new copy of the attribute object and therefor won't
mess up with other type object consuming the same role.