Don't use .perl in serialization ID creation
The source-line of 'use' statements is part of .perl output of
CompUnit::DependencySpecification, which was used to create serialization
ID's stored in .precomp files.  This adds a "serialization-id" method that
creates a shorter, but still readable string that is actually valid Raku
code, but does not contain the source-line number.  This should, possibly
drastically, reduce the number of attempts to load the same module.