Remove source-line-number from CompUnit::DependencySpecification
The source-line-number was only used for exception messages and should never
have been in the DependencySpecification object as it doesn't specify a
dependency. It also led to duplicated dependency lines in precomp files as
the dependencies are deduplicated by the .perl output of the
DependencySpecification object.

Instead, use the existing infrastructure for adding source location information
to exceptions.