Allow Str.(r)index searches on character classes
This adds 2 enums: UnicodeClass and NotUnicodeClass, and a range of
values matching the nqp::const::CCLASS_xxx constants.

It also adds some candidates to Str.index and Str.rindex, that will
dispatch on specifying a value of either of these two enums.  This
will allow you to search a string for characters (not) matching a
certain character class, e.g.

    say "foo5bar".index(NotAlphabeticChar);    # 3
    say "foo5bar".index(NotAlphanumericChar);  # Nil

This is a prove of concept.  It is implemented in core_epilogue as
an augment on the Str class, because you cannot create enums with
specific values without Pairs, and Pairs don't exist yet when Str
is being compiled.  If this is accepted, a more natural way to
implement this should be devised.