Make Rational.base at least 2.5x as fast
- completely re-imagined into a multi-method for better optimizing
- use a native string array to assemble components of the result
- much faster than 2.5x for values with more fractional digits
- introducing a :no-trailing-zeroes named parameter.

Issue GH #1872 describes that using Whatever as number of digits, may
result in an indefinite hang.  This was not fixed.  Instead, if one
wants *many* digits of a value, one has to specify that value.  But
then there is no way to get rid of trailing 0's (which the Whatever
case *does* do).  For those situations, one can now specify :no-trailing-zeroes

    (1/2).base(10,20,:no-trailing-zeroes)   # 0.5
    (2/3).base(10,20,:no-trailing-zeroes)   # 0.66666666666666666667

This probably should be emphasized as an alternative to Whatever.