Fix missing dependency information when precompiling large dependency trees
The precompilation process writes dependency information to STDOUT from which
it is read by the parent process. STDOUT is buffered by default when it's not
connected to a tty. Apparently we do not always flush the buffer on program
exit, so instead, flush every line immediately when written. There are usually
not that may anyway.
Fixes "Missing or wrong version of dependency" issue in a large proprietary
code base.