Implement `--git-cache-dir` Configure.pl option
This option makes the configure process create / update git repositories
for all of its submodules and NQP in the given folders and use those
as reference repos. The option is forwarded to a potential NQP
sub-build.

Also remove stale documentation for a not anymore implemented
`--git-reference` option. I suspect it was lost during the great build
system refactor.