Make default degree for hyper/batch make more sense
Currently, the default :degree for .hyper / .race is set to 4.  This
may be too high for small single core CPU's, and to small for bigger
beasts (such as my MBP with 2x8 cores).  Set the default to the number
of CPU cores available, and introduce the $*DEGREE dynamic variable
for setting the default for a given scope.