Make sure that Attribute.new.is_built defaults correctly
When the "is built" trait was added, all calls in the bootstrap and
the setting that created attributes, had the new "is_built" attribute
also specified.  That was in fact wrong.  What *should* have been
done, is to make sure the new "is_built" attribute would default to
the value (implicitely) given by the "has_accessor" attribute.  Only
that would ensure that Attribute's contract would not be broken.

This solves the ADT blocker (#3427)