NativeCall: add CArray[…].new(Str:D, …) candidate
This constructor creates a CArray by first encoding the given string
and then appending a NUL byte. Its primary use would be to make a copy
of a string being passed to a native function and keeping a reference
to that copy, which the `is encoded` trait does not permit.

This came up in Raku/doc#3171 and refs Raku/doc#3169.