Set $*ARGFILES to $*IN if no arguments
Not sure why it was made an IO::ArgFiles, copying over all relevant
attributes.  If there are no arguments, $*ARGFILES effectively reads
from STDIN, so not sure why special wrapping would be needed in that
case.  This is make spectest / make test clean, so if there is any
ecosystem fallout, then this commit should be reverted and tests
should be added to ensure this optimization is not attempted again.