Add mechanism for seeding typename suggestions
Before this commit:
    my String $a = "foo";
would fail with:
    Type 'String' is not declared. Did you mean 'Stringy'?
Note that this does *not* mantion 'Str', because the levenshtein
distance is too far.

After this commit it fails with:
    Type 'String' is not declared. Did you mean any of these?
        Str
        Stringy

This commit adds a generic mechanism for testing some newbie likely
wrong names, and suggest alternatives for them in addition to whatever
the levenshtein lookup will produce.  It seeds it with suggestions for
String, Integer, Float, Number and their lowercase equivalents.  Only
suggest title-cased equivalents for title-cased type names, suggest
both lowercase and title-cased for lowercase type names.