Make determining .raku / .perl switch smarter
Before this commit, if a module had a "multi method perl" in it,
or a class (other than Mu) had one, calling .raku on that class
would wrongly assume the class did *not* have a custom .perl method.
This caused breakage in roast in the 6.c-errata branch.

Now, we check whether the proto found is in Mu, and if so, if it
only has 1 dispatchee (the multi method .perl in Mu).  If it has
more than one, it implies there are more specific .perl methods in
the inheritance list, and thus the .perl method should be called,
and *not* the default .raku method.

Also extract the default logic into its own private method, for
better readability and optimizability.