Simplify leap second updating logic
Replace the Perl based "update-tai-utc.pl" script with a simple
Raku script called "add-leap-second.raku" that will take the date
for which a leap second should be added, and an optional source-
file to read.  This source-file will then be searched for markers
and the new leap second will be added at the appropriate place.

The new source-file is written to STDOUT.  An example adding a
leap second for 1 January 2021:

    $ tools/add-leap-second.raku 2021-01-01 >foo
    $ mv foo src/core.c/Rakudo/Internals.pm
    $ git diff
    @@ -811,6 +811,7 @@ implementation detail and has no serviceable parts inside"
             '2012-06-30',
             '2015-06-30',
             '2016-12-31',
    +        '2020-12-31',
             #END leap-second-dates
          );

    @@ -856,6 +857,7 @@ implementation detail and has no serviceable parts inside"
             1341100800,
             1435708800,
             1483228800,
    +        1609459200,
             #END leap-second-posix