Regex.ACCEPTS should not assume $/ is a container
Since https://github.com/rakudo/rakudo/commit/d8e859d000 an
nqp::istype is done during creation of a BUILDPLAN for a class.
In some cases, specifically when the constraint on an attribute
is a subtype, this meant calling Regex.ACCEPTS at compile time.
However, at that point in time, the nearest $/ is an immutable
Perl6::Grammar object, causing an exception to be thrown.

This commit rewrites the Regex.ACCEPTS(Any) logic to check whether
the nearest $/ is actually a container, and only sets $/ if it can.
The opportunity to rewrite using ternaries was also taken.

This fixes the breakage of e.g. the URI module in the ecosystem
as spotted by vrurg++.