Give Junction an iterator method
While methods in general autothread, some opt out of that - especially
those called by internals and that are expected to follow protocols. It
would seem `iterator` is a fair fit for that category, so implement it
as an iterator that produces 1 value when iterated: the Junction itself.
The current bad behavior in https://github.com/rakudo/rakudo/issues/3198
is eliminated by this change; it now outputs `Junction` instead.