Don't unwant the first argument of a call node
It's not clear exactly what this branch was intended to do. Removing it
resolves the bug in https://github.com/rakudo/rakudo/issues/3034, and
does not break any tests nor spectests. Its effect was to unwant - and
so sink - either the first argument of a call that itself was unwanted,
or in the case of a call node without a name, the callee. The latter was
what caused #3034. However, I can't think of any case where this is a
desirable behavior. Perhaps just a thinko?