Fix concurrency issue with re-checking a precomp file's dependencies
When the repository's identity changes (whether through changes in source files
or a `use lib` after loading modules) we re-check dependencies of precomp files
as the change may have brought in a new version of a dependency. We store the
resulting repo-id in the precomp store in the form of .repo-id files
accompanying the precomp files. When writing those files we didn't protect
against concurrent writes or reads, so the reader may have gotten an empty file
which will lead to empty $repo-ids and confused code.
Fix by locking the precomp store (protecting against concurrent writes) and
using the "rename new file over the existing file" trick (protecting against
concurrent reads).