Restore exact behaviour of List.index/indices/contains
With c9b0218544 a warning is issued with regards to using a List
as a string, when that is most likely not what one wants.  However,
that commit also changed the string on which the method operates
for a list like <a b c> from "a b c" to "abc".  This commit restores
the "a b c" behaviour.