Remove unreachable candidates
Since the nameds are optional, the candidates without nameds will
never be reached.  So they can be removed.  The alternative would
be to create candidates with required nameds, but that would lead
to a great many candidates, and since most of the work isn't done
here anyway, and one should use the method way in general anyway
if one is interested in performance, leaving it like this seems
to be the best compromise.