Make Date.Str (default formatter) about 250x as fast
It dawned on me that using sprintf (and its known overhead) was
really overkill for stringifying a Date, as we know that the values
in the object strictly have a limited value range.  So use a combo
of nqp::list_s, nqp::push_s, nqp::tostr_I and nqp::join for this
amazing performance improvement.

I guess I should really get back to looking at sprintf's
implementation again.