Rename $!storage to $!eigenstates in Junction
This has bothered me for years: $!storage is typically the name that
is being used for hash-like structures, and at least in the current
implementation of Junctions, the eigenstates are *not* stored in a
hash.  So change the name to the more descriptive '$!eigenstates`,
also to prevent false positives when searching the Rakudo code base
for '$!storage`.