Revert the $?FILE cleanup
It badly affects spectest performance.  This could have several reasons
but I suspect it is actually the extra work that CallFrame has to do.
Or it is that it somehow causes test libraries to have to be recompiled
again and again.  Anyways, it was a hack, and we need a better way to
convey what the current file is cleanly, and how to provide better
support for execuction errors from CUR::I installed files.