Fix mixing in Raku-level code to the grammar
The grammar is written in NQP, and so wants its match objects to have
an nqp::list for quantified captures. An internals change resulted in us
always creating a Raku Array instead; this is fine for everything except
when we want to mix into the Raku grammar itself. In that case, we could
end up with quantified things only collecting the last captured value,
since the NQP Match construction code didn't understand it was a list.
To fix this, we retain the NQP behavior under one method name, and the
Raku behavior under a differnet method name.

Fixes https://github.com/rakudo/rakudo/issues/3345.