Avoid re-resolving the same dependencies multiple times
When the repository chain's id changes (e.g. because a source file in a -Ilib
repository got edited) we need to re-resolve the dependencies of the precomp
files we load as the change could have brought in a new prefered version of a
dependency. However doing so once is quite sufficient. There's no need to
re-resolve the same dependency specification over and over which could happen
in large applications with many modules sharing similar dependencies.
Brings down compilation time in a real life code base from 110 to 14 seconds.