Make Backtrace independent of context.
Because `Backtrace` depends on expections to obtain low-level list of
frames, its output is dependent on `Exception` implementation and
optimizations. For example, _S32-exceptions/misc.t_ test contains the
following note: "the number of blocks (2) below is important; otherwise
the $bt.list.elems will give the wrong number". But, in fact, not only
it depends on the number of blocks (it would not anymore), but on where
these blocks are located! Moving them at the beginning of the test
results in failures with number of frames returns being 6 instead of
expected 4. This happens because not enough stats were collected to
inline method `throw` of `Exception`. This is just is my guess, but it
looks quite reasonable.

It is also possible that a change in `throw` implementation could make
it uninlinable causing the same breakage.

This patch explicitly makes `Backtrace` skipping all frames up to and
including those with 'Backtrace.pm6' file name. Note that for speed it
is not tested wether what is located is really a file name. The only
case where this might result in wrong results is if `Exception::throw`
would be wrapped with a code coming from a path containing
`Backtrace.pm6` substring.

IMPORTANT The correct implementation of Backtrace requires
`nqp::backtrace` accept not only exceptions but also contexts. With
non-inlinable `Backtrace::new` it would guarantee that skippin 1 frame
is sufficient for correct work.