The signal Supply should not claim sanity
That implies that it is serial (since sanity is a stronger promise than
serialized). However, it's not serial. This meant that we could end up
with concurrent execution in a situation like:

    react whenever signal(SIGINT) { #`(long running stuff) }

Because the one-whenever form of `react` doesn't add a lock, but instead
calls .sanitize on the input `Supply`, and that was wrongly being a
no-op because signal's `Supply` claimed it was already fine. This
addresses #3547.