Save upto 2 allocations per Date/DateTime.yyyy-mm-dd
- and dd-mm-yyyy and mm-dd-yyyy friends
- 1 Int allocation if month > 9
- 1 Int allocation if day > 9

For some reason, *in the core setting*, having a native int attribute
as part of a nqp::push_s (which should just convert it to a native
string), gives strange errors.  The workaround was to use nqp::to_str_I
on a nqp::getattr_i, which caused the Int allocation.  Oddly enough
using an native int attribute as part of a nqp::concat just works, so
use that together with an nqp::x in which the repeat factor is determined
by the result of the condition < 10 (which yields either 1 or 0).