Re-instate &GENERATE-ROLE-FROM-VALUE again
This caused unexpected breakage in the MinG module at precompilation,
complaining there was no REPR for BOOTContext objects.  So the
assumption is that the method call somehow added some context into
closures being precompiled that were unexpected.

This reverts commit b639221dd7eb3bfa2ff1c5cdf37a4e50628ea98e, but also
marked the sub as "implementation-detail", so it won't show up in any
introspection by default.