Streamlining Rakudo::SEQUENCE.iterator (8/8)
Add some more binding that were missed before.

This is the last of my attempts to make SEQUENCE (aka ... handling)
faster.  The original code predates Rakudo Iterator technology and
is in bad need of a complete redesign: SEQUENCE should produce an
iterator, and should handle the simple cases with a simple iterator.
Using gather / take is just too much overhead for what, in most cases,
can be a simple iterator, with possible local optimizations.

Take for example the general "a" ... "z" unicode case: it will take
the .ord of the previous value, add one, and change that to a .chr
again.  An iterator would keep the codepoint value, and produce a
.chr out of that, so no .ord taking for each produced value would
be needed.