Add an In Memoriam file to complement CREDITS
While I was going through the CREDITS file the other day, I couldn't
help but notice that 2 people in there had passed away.  So I thought
it would be nice to commemorate them more extensively in a separate
file, to be able to credit their work better, and keep their memories
alive.

Please note that this by no means is intended to be the final version
of this file.  If you feel like you have something to add to it, or
feel that corrections should be made, then please open up an issue,
or create a Pull Request for it.  Thank you!