Give all regexes their own $/
We would elide this for forms made from quote constructs. It's unclear
why we did this; there are no spectests whatsoever that depend on such
behavior and we'd clobber what was in the $/ in most cases anyway. The
real issue this caused, however, was in `m:g/... {}/`, where the `{}`
causes a match object sync into `$/`. When it used the outer one, it
would end up overwriting the `$/` containing a `List` from the `m:g`
with the match object made when syncing `$/` up for the code block,
thus resulting in the bad behavior observed in #3554. Giving every regex
its own $/ lexical solves this, without spectest fallout. It's also a
simplification in the compiler.