Remove undocumented Seq.new-consumed
And replace it with a Seq.new() candidate.  Its existence was only
needed to create an EVALlable representation of a Seq of which the
iterator had already been taken.  This created a documentation
issue, as the method was *not* an implementation detail.  Yet
documenting it would only cause more questions, as it strayed very
far from the norm.

This needs to have some 6.d spectest changes, but fortunately no
errata fixes.

Historically, this was probably done to make sure that Seq.new was
not a multi.  But with the current state of optimizations, that
should no longer be an issue.  However, if this *does* turn out to
be a performance issue, then just revert this commit.