Adopt MoarVM-provided value_desc_cont
This gets rid of the Rakudo Scalar container descriptor, which was a big
chunk of the extops library (which we're gradually aiming to get rid
of). It also removes an assumption on P6opaque and Scalar layout, which
makes things a bit more robust. Curiously, on a benchmark I did to check
the consequence of this, it seems that we may also gain a little speed
too (I saw a ~1% CPU cycle reduction on the thing I measured). That goes
slightly against my intuition, but I won't complain.

Moving this was planned back when I did the last round of Scalar
container changes, but I never got around to it. However, it will ease
optimizing container access in arg guards to have it in MoarVM, so the
time came when there was a decent motivation to do it. I've tried to
keep the JVM and JS backends working after this too; there's not so much
to win by moving the code in those, however.