Make the 'is' routine in Test not throwing an error on failure when the type can't be stringified, falling back on .raku instead.
A good example is the Buf type that has a Str routine that fail to encourage the user to use encode instead.

is Buf.new(42), Buf.new(43)

This will fail with the error :
Cannot use a Buf as a string, but you called the Stringy method on it
instead of the regular got, expected message.

This fix issue #3535