Add an iterator multi candidate to Supply
A problem with Seq and List supply coercions was that take could be used
inside of them, which would mix values taken with those that are
actually emitted by the supply in the coercion. Instead of using gather
to handle coercions in the first place, this adds an iterator method to
Supply to handle that for these two methods. Because of this,
Supply.list no longer needs to create a throwaway Seq to coerce to a
List.