Make Date.succ faster for days 28-31
If the $!day in a Date object is < 28, then we can be sure that we
just need to increment $!day to get the next Date.  However, for
days 28-31 it would just take the full daycount + days path, which
is pretty expensive as it involves a lot of calculations.  After
this commit, days 28-31 will check the (lot less expensive) number
of days in the current month, and if within range, take the fast
path as well.