Check reified before sending it off to R:I
Turns out, something in the bowels of the module loader is creating
an empty String object without setting the $!reified attribute, and
calls .reverse on it.  In the old code, this would be caught by
the conditional: the new code didn't do that, assuming it would be
handled inside R:I.  Restored the old situation, but still calling
the new ReifiedListReverse iterator.

This fixes test failures as seen in CircleCI and friends.